refactor: split SportsSidebarMenu into focused sub-components
